aws-docker/mem0_add_message/image/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.4.1
greenlet==3.1.1
grpcio==1.66.2
grpcio-tools==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-text-splitters==0.2.4
langsmith==0.1.129
marshmallow==3.22.0
mem0ai==0.1.16
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.25.0
numpy==1.26.4
openai==1.50.2
orjson==3.10.7
packaging==24.1
portalocker==2.10.1
posthog==3.6.6
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.11.3
rank-bm25==0.2.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.13.1
pandas
datetime